fix: add a separate module command_handlers.py
